MakeMyTripFramework/requirement.txt
allure-behave==2.9.45

allure-pytest==2.8.40

allure-python-commons==2.9.45

apipkg==1.5

astor==0.8.1

async-generator==1.10

atomicwrites==1.4.0

attrs==20.3.0

behave==1.2.6

certifi==2020.12.5

cffi==1.15.0

chardet==4.0.0

colorama==0.4.4

configparser==5.0.2


iniconfig==1.1.1

openpyxl==3.0.7

outcome==1.1.0

packaging==20.9

parse==1.19.0

parse-type==0.5.2

pluggy==0.13.1

py==1.10.0

pycparser==2.21

pyOpenSSL==21.0.0

pyparsing==2.4.7

pytest==6.2.3

pytest-ast-transformer==1.0.3

pytest-forked==1.3.0

pytest-html==3.1.1

pytest-metadata==1.11.0

pytest-soft-assertions==0.1.2

pytest-xdist==2.2.1

requests==2.25.1

selenium==4.0.0

six==1.15.0

sniffio==1.2.0

sortedcontainers==2.4.0

texttable==1.6.3

toml==0.10.2

trio==0.19.0

trio-websocket==0.9.2

urllib3==1.26.4

webdriver-manager==3.3.0

wsproto==1.0.0

